refactor: Extract options to a separate file (#15325) (CP: 23.3) (#15444)
