+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/NodeTasksHillaTest.java
@@ -18,7 +18,6 @@ package com.vaadin.flow.server.frontend;
 import com.vaadin.experimental.FeatureFlags;
 import com.vaadin.flow.di.Lookup;
 import com.vaadin.flow.server.ExecutionFailedException;
-import com.vaadin.flow.server.frontend.NodeTasks.Options;
 import com.vaadin.flow.server.frontend.scanner.ClassFinder;
 import com.vaadin.flow.server.frontend.scanner.ClassFinder.DefaultClassFinder;
 import org.apache.commons.io.FileUtils;
