+++ flow-server/src/test/java/com/vaadin/flow/server/frontend/TaskUpdateViteTest.java
@@ -88,8 +88,8 @@ public class TaskUpdateViteTest {
 
         String template = IOUtils.toString(generatedConfigFile.toURI(),
                 StandardCharsets.UTF_8);
-        NodeTasks.Options builder = new NodeTasks.Options(
-                Mockito.mock(Lookup.class), temporaryFolder.getRoot(),
+        Options builder = new Options(Mockito.mock(Lookup.class),
+                temporaryFolder.getRoot(),
                 temporaryFolder.newFolder("generated"),
                 temporaryFolder.newFolder("frontend"), "target");
         builder.withJarFrontendResourcesFolder(
